+++ src/dev/ukanth/ufirewall/plugin/FireReceiver.java
@@ -17,6 +17,8 @@ import android.content.Context;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.os.Bundle;
+import android.os.Handler;
+import android.os.Message;
 import android.preference.PreferenceManager;
 import android.widget.Toast;
 import dev.ukanth.ufirewall.Api;
@@ -57,7 +59,15 @@ public final class FireReceiver extends BroadcastReceiver
         BundleScrubber.scrub(intent);
         BundleScrubber.scrub(intent.getBundleExtra(com.twofortyfouram.locale.Intent.EXTRA_BUNDLE));
         final Bundle bundle = intent.getBundleExtra(com.twofortyfouram.locale.Intent.EXTRA_BUNDLE);
+        final SharedPreferences prefs2 = context.getSharedPreferences(Api.PREFS_NAME, 0);
+        final String oldPwd = prefs2.getString(Api.PREF_PASSWORD, "");
+		final String newPwd = context.getSharedPreferences(Api.PREF_FIREWALL_STATUS, 0).getString("LockPassword", "");
 
+		final Handler toaster = new Handler() {
+			public void handleMessage(Message msg) {
+				if (msg.arg1 != 0)Toast.makeText(context, msg.arg1, Toast.LENGTH_SHORT).show();
+			}
+		};
         /*
          * Final verification of the plug-in Bundle before firing the setting.
          */
@@ -68,37 +78,77 @@ public final class FireReceiver extends BroadcastReceiver
         	final boolean multimode = prefs.getBoolean("enableMultiProfile", false);
         	final boolean disableToasts = prefs.getBoolean("disableTaskerToast", false);
     		SharedPreferences.Editor editor = prefs.edit();
-        	if(index != null && multimode){
+    		final Message msg = new Message();
+        	if(index != null){
         		int id = Integer.parseInt(index);
-        		editor.putInt("storedPosition", id);
-       			editor.commit();
         		switch(id){
         		case 0:
+        			if(applyRules(context,msg,toaster)){
+						Api.setEnabled(context, true, false);
+					}
+       				break;
+       				
+        		case 1:	
+        			if(oldPwd.length() == 0 && newPwd.length() == 0){
+						if (Api.purgeIptables(context, false)) {
+							msg.arg1 = R.string.toast_disabled;
+							toaster.sendMessage(msg);
+							Api.setEnabled(context, false, false);
+						} else {
+							msg.arg1 = R.string.toast_error_disabling;
+							toaster.sendMessage(msg);
+						}
+					} else {
+						msg.arg1 = R.string.widget_disable_fail;
+						toaster.sendMessage(msg);
+					}
+        			break;
+        		case 2:
     				Api.PREFS_NAME = "AFWallPrefs";
     				break;
-    			case 1:
+    			case 3:
     				Api.PREFS_NAME = "AFWallProfile1";
     				break;
-    			case 2:
+    			case 4:
     				Api.PREFS_NAME = "AFWallProfile2";
     				break;
-    			case 3:
+    			case 5:
     				Api.PREFS_NAME = "AFWallProfile3";
     				break;
     			default:
     				break;
 				}
-        		if (Api.isEnabled(context)) {
-        			if(!disableToasts){
-        				Toast.makeText(context, R.string.tasker_apply, Toast.LENGTH_SHORT).show();	
+        		if(id > 1){
+        			editor.putInt("storedPosition", (id-2));
+           			editor.commit();
+        			if(multimode) {
+        				if (Api.isEnabled(context)) {
+                			if(!disableToasts){
+                				Toast.makeText(context, R.string.tasker_apply, Toast.LENGTH_SHORT).show();	
+                			}
+                			Api.applySavedIptablesRules(context, true);
+                		} else {
+               				Toast.makeText(context, R.string.tasker_disabled, Toast.LENGTH_SHORT).show();
+                		}		
         			}
-        			Api.applySavedIptablesRules(context, true);
-        		} else {
-       				Toast.makeText(context, R.string.tasker_disabled, Toast.LENGTH_SHORT).show();
+        			else {
+                		Toast.makeText(context, R.string.tasker_muliprofile, Toast.LENGTH_LONG).show();
+                	}
         		}
-        	} else {
-        		Toast.makeText(context, R.string.tasker_muliprofile, Toast.LENGTH_LONG).show();
-        	}
+        	} 
         }
     }
+    
+    private boolean applyRules(Context context,Message msg, Handler toaster) {
+		boolean success = false;
+		if (Api.applySavedIptablesRules(context, false)) {
+			msg.arg1 = R.string.toast_enabled;
+			toaster.sendMessage(msg);
+			success = true;
+		} else {
+			msg.arg1 = R.string.toast_error_enabling;
+			toaster.sendMessage(msg);
+		}
+		return success;
+	}
 }